+++ contrib/opencensus-ext-azure/opencensus/ext/azure/log_exporter/__init__.py
@@ -77,7 +77,7 @@ class Worker(threading.Thread):
         )
 
     def run(self):
-        # Indicate that this thread is an exporter thread. Used for auto-collection.
+        # Indicate that this thread is an exporter thread.
         execution_context.set_is_exporter_thread(True)
         src = self._src
         dst = self._dst