+++ contrib/opencensus-ext-azure/tests/test_azure_trace_exporter.py
@@ -45,7 +45,7 @@ class TestAzureExporter(unittest.TestCase):
         self.assertRaises(ValueError, lambda: trace_exporter.AzureExporter())
         Options._default.instrumentation_key = instrumentation_key
 
-    @mock.patch('requests.Session.post', return_value=mock.Mock())
+    @mock.patch('requests.post', return_value=mock.Mock())
     def test_emit_empty(self, request_mock):
         exporter = trace_exporter.AzureExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
@@ -225,7 +225,7 @@ class TestAzureExporter(unittest.TestCase):
             envelope.data.baseType,
             'RemoteDependencyData')
 
-        # SpanKind.SERVER HTTP
+        # SpanKind.SERVER HTTP - 200 request
         envelope = exporter.span_data_to_envelope(SpanData(
             name='test',
             context=SpanContext(
@@ -239,6 +239,8 @@ class TestAzureExporter(unittest.TestCase):
             parent_span_id='6e0c63257de34c93',
             attributes={
                 'http.method': 'GET',
+                'http.path': '/wiki/Rabbit',
+                'http.route': '/wiki/Rabbit',
                 'http.url': 'https://www.wikipedia.org/wiki/Rabbit',
                 'http.status_code': 200,
             },
@@ -266,6 +268,9 @@ class TestAzureExporter(unittest.TestCase):
             envelope.tags['ai.operation.id'],
             '6e0c63257de34c90bf9efcd03927272e')
         self.assertEqual(
+            envelope.tags['ai.operation.name'],
+            'GET /wiki/Rabbit')
+        self.assertEqual(
             envelope.time,
             '2010-10-24T07:28:38.123456Z')
         self.assertEqual(
@@ -279,7 +284,80 @@ class TestAzureExporter(unittest.TestCase):
             '200')
         self.assertEqual(
             envelope.data.baseData.name,
-            'GET https://www.wikipedia.org/wiki/Rabbit')
+            'GET /wiki/Rabbit')
+        self.assertEqual(
+            envelope.data.baseData.success,
+            True)
+        self.assertEqual(
+            envelope.data.baseData.url,
+            'https://www.wikipedia.org/wiki/Rabbit')
+        self.assertEqual(
+            envelope.data.baseType,
+            'RequestData')
+
+        # SpanKind.SERVER HTTP - Failed request
+        envelope = exporter.span_data_to_envelope(SpanData(
+            name='test',
+            context=SpanContext(
+                trace_id='6e0c63257de34c90bf9efcd03927272e',
+                span_id='6e0c63257de34c91',
+                trace_options=TraceOptions('1'),
+                tracestate=Tracestate(),
+                from_header=False,
+            ),
+            span_id='6e0c63257de34c92',
+            parent_span_id='6e0c63257de34c93',
+            attributes={
+                'http.method': 'GET',
+                'http.path': '/wiki/Rabbit',
+                'http.route': '/wiki/Rabbit',
+                'http.url': 'https://www.wikipedia.org/wiki/Rabbit',
+                'http.status_code': 400,
+            },
+            start_time='2010-10-24T07:28:38.123456Z',
+            end_time='2010-10-24T07:28:38.234567Z',
+            stack_trace=None,
+            links=None,
+            status=None,
+            annotations=None,
+            message_events=None,
+            same_process_as_parent_span=None,
+            child_span_count=None,
+            span_kind=SpanKind.SERVER,
+        ))
+        self.assertEqual(
+            envelope.iKey,
+            '12345678-1234-5678-abcd-12345678abcd')
+        self.assertEqual(
+            envelope.name,
+            'Microsoft.ApplicationInsights.Request')
+        self.assertEqual(
+            envelope.tags['ai.operation.parentId'],
+            '|6e0c63257de34c90bf9efcd03927272e.6e0c63257de34c93.')
+        self.assertEqual(
+            envelope.tags['ai.operation.id'],
+            '6e0c63257de34c90bf9efcd03927272e')
+        self.assertEqual(
+            envelope.tags['ai.operation.name'],
+            'GET /wiki/Rabbit')
+        self.assertEqual(
+            envelope.time,
+            '2010-10-24T07:28:38.123456Z')
+        self.assertEqual(
+            envelope.data.baseData.id,
+            '|6e0c63257de34c90bf9efcd03927272e.6e0c63257de34c92.')
+        self.assertEqual(
+            envelope.data.baseData.duration,
+            '0.00:00:00.111')
+        self.assertEqual(
+            envelope.data.baseData.responseCode,
+            '400')
+        self.assertEqual(
+            envelope.data.baseData.name,
+            'GET /wiki/Rabbit')
+        self.assertEqual(
+            envelope.data.baseData.success,
+            False)
         self.assertEqual(
             envelope.data.baseData.url,
             'https://www.wikipedia.org/wiki/Rabbit')
@@ -415,7 +493,7 @@ class TestAzureExporter(unittest.TestCase):
         self.assertEqual(len(os.listdir(exporter.storage.path)), 1)
         exporter._stop()
 
-    @mock.patch('requests.Session.post', return_value=mock.Mock())
+    @mock.patch('requests.post', return_value=mock.Mock())
     def test_transmission_lease_failure(self, requests_mock):
         requests_mock.return_value = MockResponse(200, 'unknown')
         exporter = trace_exporter.AzureExporter(
@@ -435,7 +513,7 @@ class TestAzureExporter(unittest.TestCase):
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
         exporter.storage.put([1, 2, 3])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(200, None)
             del post.return_value.text
             exporter._transmit_from_storage()
@@ -450,7 +528,7 @@ class TestAzureExporter(unittest.TestCase):
         )
         exporter.storage.put([1, 2, 3])
         exporter.storage.put([1, 2, 3])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(200, 'unknown')
             exporter._transmit_from_storage()
         self.assertIsNone(exporter.storage.get())
@@ -463,7 +541,7 @@ class TestAzureExporter(unittest.TestCase):
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
         exporter.storage.put([1, 2, 3])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(206, 'unknown')
             exporter._transmit_from_storage()
         self.assertIsNone(exporter.storage.get())
@@ -476,7 +554,7 @@ class TestAzureExporter(unittest.TestCase):
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
         exporter.storage.put([1, 2, 3, 4, 5])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(206, json.dumps({
                 'itemsReceived': 5,
                 'itemsAccepted': 3,
@@ -504,7 +582,7 @@ class TestAzureExporter(unittest.TestCase):
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
         exporter.storage.put([1, 2, 3])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(206, json.dumps({
                 'itemsReceived': 3,
                 'itemsAccepted': 2,
@@ -526,7 +604,7 @@ class TestAzureExporter(unittest.TestCase):
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
         exporter.storage.put([1, 2, 3, 4, 5])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(206, json.dumps({
                 'itemsReceived': 5,
                 'itemsAccepted': 3,
@@ -548,7 +626,7 @@ class TestAzureExporter(unittest.TestCase):
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
         exporter.storage.put([1, 2, 3])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(400, '{}')
             exporter._transmit_from_storage()
         self.assertEqual(len(os.listdir(exporter.storage.path)), 0)
@@ -560,7 +638,7 @@ class TestAzureExporter(unittest.TestCase):
             storage_path=os.path.join(TEST_FOLDER, self.id()),
         )
         exporter.storage.put([1, 2, 3])
-        with mock.patch('requests.Session.post') as post:
+        with mock.patch('requests.post') as post:
             post.return_value = MockResponse(500, '{}')
             exporter._transmit_from_storage()
         self.assertIsNone(exporter.storage.get())