+++ contrib/opencensus-ext-httplib/opencensus/ext/httplib/trace.py
@@ -61,14 +61,10 @@ def wrap_httplib_request(request_func):
     """
 
     def call(self, method, url, body, headers, *args, **kwargs):
-        try:
-            # Check if request was sent from an exporter. If so, do not wrap.
-            if execution_context.is_exporter_thread():
-                return request_func(self, method, url, body,
+        # Check if request was sent from an exporter. If so, do not wrap.
+        if execution_context.is_exporter_thread():
+            return request_func(self, method, url, body,
                                 headers, *args, **kwargs)
-        except AttributeError:
-            # If not set, continue with wrapping
-            pass
         _tracer = execution_context.get_opencensus_tracer()
         blacklist_hostnames = execution_context.get_opencensus_attr(
             'blacklist_hostnames')
@@ -108,13 +104,9 @@ def wrap_httplib_response(response_func):
     """
 
     def call(self, *args, **kwargs):
-        try:
-            # Check if request was sent from an exporter. If so, do not wrap.
-            if execution_context.is_exporter_thread():
-                return response_func(self, *args, **kwargs)
-        except AttributeError:
-            # If not set, continue with wrapping
-            pass
+        # Check if request was sent from an exporter. If so, do not wrap.
+        if execution_context.is_exporter_thread():
+            return response_func(self, *args, **kwargs)
         _tracer = execution_context.get_opencensus_tracer()
         current_span_id = execution_context.get_opencensus_attr(
             'httplib/current_span_id')