+++ contrib/opencensus-ext-requests/opencensus/ext/requests/trace.py
@@ -61,13 +61,9 @@ def trace_integration(tracer=None):
 def wrap_requests(requests_func):
     """Wrap the requests function to trace it."""
     def call(url, *args, **kwargs):
-        try:
-            # Check if request was sent from an exporter. If so, do not wrap.
-            if execution_context.is_exporter_thread():
-                return requests_func(url, *args, **kwargs)
-        except AttributeError:
-            # If not set, continue with wrapping
-            pass
+        # Check if request was sent from an exporter. If so, do not wrap.
+        if execution_context.is_exporter_thread():
+            return requests_func(url, *args, **kwargs)
         blacklist_hostnames = execution_context.get_opencensus_attr(
             'blacklist_hostnames')
         parsed_url = urlparse(url)
@@ -100,13 +96,9 @@ def wrap_requests(requests_func):
 
 def wrap_session_request(wrapped, instance, args, kwargs):
     """Wrap the session function to trace it."""
-    try:
-        # Check if request was sent from an exporter. If so, do not wrap.
-        if execution_context.is_exporter_thread():
-            return wrapped(*args, **kwargs)
-    except AttributeError:
-        # If not set, continue with wrapping
-        pass
+    # Check if request was sent from an exporter. If so, do not wrap.
+    if execution_context.is_exporter_thread():
+        return wrapped(*args, **kwargs)
 
     method = kwargs.get('method') or args[0]
     url = kwargs.get('url') or args[1]